@@ -41,94 +42,105 @@
     "hot reload",
     "react-hot-loader",
     "enzyme",
-    "mocha",
-    "chai"
+    "jest"
   ],
   "license": "MIT",
   "devDependencies": {
     "@types/base16": "^1.0.1",
-    "@types/chai": "^4.0.0",
-    "@types/enzyme": "^2.7.1",
+    "@types/enzyme": "^2.8.1",
     "@types/isomorphic-fetch": "0.0.34",
-    "@types/mocha": "^2.2.38",
-    "@types/node": "^7.0.0",
-    "@types/react": "^15.0.22",
-    "@types/react-addons-test-utils": "^0.14.16",
-    "@types/react-dom": "^15.5.0",
-    "@types/react-helmet": "^5.0.0",
-    "@types/react-redux": "^4.4.38",
-    "@types/react-router": "^3.0.3",
-    "@types/react-router-redux": "^4.0.4",
+    "@types/jest": "^19.2.3",
+    "@types/node": "^8.0.2",
+    "@types/prop-types": "^15.5.1",
+    "@types/raven-js": "^3.10.0",
+    "@types/react": "^15.0.31",
+    "@types/react-addons-test-utils": "^0.14.19",
+    "@types/react-dom": "^15.5.1",
+    "@types/react-helmet": "^5.0.3",
+    "@types/react-intl": "^2.3.0",
+    "@types/react-redux": "^4.4.45",
     "@types/redux": "^3.6.31",
-    "@types/redux-devtools": "^3.0.36",
-    "@types/redux-devtools-dock-monitor": "^1.1.29",
-    "@types/redux-devtools-log-monitor": "^1.0.29",
-    "@types/redux-mock-store": "0.0.7",
-    "@types/redux-thunk": "^2.1.31",
-    "@types/webpack": "^2.2.13",
+    "@types/redux-devtools": "^3.0.38",
+    "@types/redux-devtools-dock-monitor": "^1.1.31",
+    "@types/redux-devtools-log-monitor": "^1.0.31",
+    "@types/redux-form": "^6.6.3",
+    "@types/webpack": "^3.0.0",
     "awesome-typescript-loader": "^3.0.0-beta.18",
     "babel-core": "^6.22.1",
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.22.0",
-    "chai": "^3.5.0",
-    "chai-enzyme": "^0.7.0",
-    "cross-env": "^4.0.0",
-    "electron": "^1.4.15",
-    "enzyme": "^2.8.2",
-    "fetch-mock": "^5.8.1",
+    "cross-env": "^5.0.1",
+    "enzyme": "^2.9.1",
     "file-loader": "^0.11.1",
-    "husky": "^0.13.2",
-    "karma": "^1.4.0",
-    "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-es6-shim": "^1.0.0",
-    "karma-firefox-launcher": "^1.0.1",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.3",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-typescript": "^3.0.0",
-    "karma-webpack": "^2.0.3",
-    "mocha": "^3.2.0",
-    "phantomjs-prebuilt": "^2.1.14",
+    "husky": "^0.14.0",
+    "jest": "^20.0.3",
+    "jest-enzyme": "3.2.0",
     "react-addons-test-utils": "~15.6.0",
-    "react-hot-loader": "^1.3.1",
+    "react-hot-loader": "^3.0.0-beta.7",
     "react-test-renderer": "^15.5.4",
-    "redux-mock-store": "^1.2.1",
     "rimraf": "^2.5.4",
+    "ts-jest": "^20.0.4",
     "tslint": "^4.3.1",
     "tslint-loader": "^3.3.0",
     "tslint-microsoft-contrib": "^4.0.1",
     "tslint-react": "^3.0.0",
-    "typescript": "2.3.0",
+    "typescript": "^2.4.0",
     "url-loader": "^0.5.7",
     "validate-commit-msg": "^2.12.1",
     "wallaby-webpack": "0.0.38",
     "webpack": "^2.3.0",
-    "webpack-dev-middleware": "^1.10.1",
+    "webpack-dev-middleware": "^1.11.0",
     "webpack-dev-server": "^2.2.0",
     "webpack-hot-middleware": "^2.15.0",
     "webpack-manifest-plugin": "^1.1.0"
   },
   "dependencies": {
-    "chalk": "^1.1.3",
-    "csstips": "0.2.0",
+    "chalk": "^2.0.1",
+    "csstips": "^0.2.2",
     "csx": "^8.2.1",
     "es6-promise": "^4.0.5",
     "express": "^4.14.0",
     "isomorphic-fetch": "^2.2.1",
+    "prop-types": "^15.5.10",
+    "raven-for-redux": "^1.0.0",
+    "raven-js": "^3.16.0",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "react-helmet": "^5.0.2",
+    "react-intl": "^2.3.0",
     "react-redux": "^5.0.2",
-    "react-router": "^3.0.2",
-    "react-router-redux": "^4.0.7",
+    "react-router5": "^5.0.2",
     "redux": "^3.6.0",
-    "redux-connect": "^5.1.0",
+    "redux-form": "^7.0.0",
     "redux-logger": "^3.0.0",
-    "redux-thunk": "^2.2.0",
+    "redux-router5": "^5.0.0",
+    "redux-saga": "^0.15.4",
+    "router5": "^5.1.0",
     "serve-favicon": "^2.3.2",
     "typestyle": "^1.0.0"
+  },
+  "jest": {
+    "mapCoverage": true,
+    "transform": {
+      "^.+\\.tsx?$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "cacheDirectory": "<rootDir>/.jest/cache",
+    "collectCoverage": true,
+    "coverageDirectory": "<rootDir>/coverage",
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js"
+    },
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "json"
+    ],
+    "moduleDirectories": [
+      "node_modules",
+      "./"
+    ],
+    "setupTestFrameworkScriptFile": "./node_modules/jest-enzyme/lib/index.js"
   }
 }
