@@ -6,16 +6,16 @@ var config = {
   port: process.env.PORT || 8889,
   karmaPort: 9876,
   ssr: true,
-
+  sentry: false,
   // This part goes to React-Helmet for Head of our HTML
   app: {
     head: {
       title: 'Crazy Factory',
       meta: [
-        { charset: 'utf-8' },
-        { 'http-equiv': 'x-ua-compatible', content: 'ie=edge' },
-        { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-        { name: 'description', content: 'React Redux Typescript' },
+        {charset: 'utf-8'},
+        {'http-equiv': 'x-ua-compatible', content: 'ie=edge'},
+        {name: 'viewport', content: 'width=device-width, initial-scale=1'},
+        {name: 'description', content: 'React Redux Typescript'},
       ]
     }
   }
