@@ -1,7 +1,10 @@
 'use strict';
 
-if (process.env.NODE_ENV === 'production') {
+if (process.env.NODE_ENV === 'production')
+{
   module.exports = require('./prod');
-} else {
+}
+else
+{
   module.exports = require('./dev');
 }
