@@ -1,5 +1,6 @@
 var path = require('path');
 var fs = require('fs');
+var utils = require('./utils');
 
 var nodeModules = {};
 fs.readdirSync('node_modules')
@@ -29,7 +30,8 @@ var config = {
   },
 
   module: {
-    loaders: [{
+    loaders: [
+      {
         test: /\.(jpe?g|png|gif)$/i,
         loader: 'url-loader?limit=1000&name=images/[hash].[ext]'
       },
@@ -65,22 +67,9 @@ var config = {
   }
 };
 
-const copySync = (src, dest, overwrite) => {
-  if (overwrite && fs.existsSync(dest)) {
-    fs.unlinkSync(dest);
-  }
-  const data = fs.readFileSync(src);
-  fs.writeFileSync(dest, data);
-}
-
-const createIfDoesntExist = dest => {
-  if (!fs.existsSync(dest)) {
-    fs.mkdirSync(dest);
-  }
-}
-
-createIfDoesntExist('./build');
-createIfDoesntExist('./build/public');
-copySync('./src/favicon.ico', './build/public/favicon.ico', true);
+utils.copySyncIfDoesntExist('./config/main.js', './config/main.local.js');
+utils.createIfDoesntExist('./build');
+utils.createIfDoesntExist('./build/public');
+utils.copySync('./src/favicon.ico', './build/public/favicon.ico', true);
 
 module.exports = config;
