@@ -1,13 +1,13 @@
 const SiteStyles = {
-    backgroundColor: {
-        backgroundColor: "lightblue"
-    },
-    linkColor: {
-        color: "darkslateblue"
-    },
-    linkColorOnHovered: {
-        color: "deepskyblue"
-    }
+  backgroundColor: {
+    backgroundColor: "lightblue"
+  },
+  linkColor: {
+    color: "darkslateblue"
+  },
+  linkColorOnHovered: {
+    color: "deepskyblue"
+  }
 };
 
-export { SiteStyles };
+export {SiteStyles};
