@@ -1,34 +1,36 @@
 import * as React from "react";
-import { Link } from "react-router";
-import { style } from "typestyle";
+import {FormattedMessage} from "react-intl";
+import {style} from "typestyle";
+import {Link} from "./Link";
 
-const Styles = {
-    nav: style({
+const styles = {
+  nav: style({
+    $nest: {
+      ul: {
         $nest: {
-            ul: {
-                $nest: {
-                    li: {
-                        color: "blue",
-                        display: "inline",
-                        padding: "5px"
-                    }
-                },
-                listStyleType: "none",
-                padding: 0
-            }
-        }
-    })
+          li: {
+            color: "blue",
+            display: "inline",
+            padding: "5px"
+          }
+        },
+        listStyleType: "none",
+        padding: 0
+      }
+    }
+  })
 };
 
 const Header = () => (
-  <nav className={Styles.nav}>
+  <nav className={styles.nav}>
     <ul>
-      <li><Link to="/">Home</Link></li>
-      <li><Link to="about">About</Link></li>
-      <li><Link to="counter">Counter</Link></li>
-      <li><Link to="stars">Stars</Link></li>
+      <li><Link name="home"><FormattedMessage id="home" defaultMessage="Home" /></Link></li>
+      <li><Link name="about"><FormattedMessage id="about" defaultMessage="About Us" /></Link></li>
+      <li><Link name="counter"><FormattedMessage id="counter" defaultMessage="Counter" /></Link></li>
+      <li><Link name="stars"><FormattedMessage id="stars" defaultMessage="Stars" /></Link></li>
+      <li><Link name="register"><FormattedMessage id="register" defaultMessage="Register" /></Link></li>
     </ul>
   </nav>
 );
 
-export { Header, Styles };
+export {Header};
