@@ -1,1 +1,2 @@
-export { Header } from "./Header";
+export {Header} from "./Header";
+export {RegisterReduxForm} from "./RegisterForm";
