@@ -1,7 +1,8 @@
 /** Exporting Containers for Easier Imports */
-export { Html } from "./Html";
-export { App } from "./App";
-export { Home } from "./Home";
-export { About } from "./About";
-export { Counter } from "./Counter";
-export { Stars } from "./Stars";
+export {Html} from "./Html";
+export {App} from "./App";
+export {HomePage} from "./HomePage";
+export {AboutPage} from "./AboutPage";
+export {CounterPage} from "./CounterPage";
+export {StarsPage} from "./StarsPage";
+export {RegisterPage} from "./RegisterPage";
