@@ -1,83 +1,34 @@
-import { IStars, IStarsAction } from "models/starsModel";
+import promiseReducer, {IRequestType} from "../../helpers/promiseReducer";
+import {IAction, IState} from "./baseModule";
 
-/** Action Types */
-export const GET_REQUEST: string = "stars/GET_REQUEST";
-export const GET_SUCCESS: string = "stars/GET_SUCCESS";
-export const GET_FAILURE: string = "stars/GET_FAILURE";
-
-/** Initial State */
-const initialState: IStars = {
-  isFetching: false
+export const LOAD_STARS: string = "stars/LOAD_STARS";
+export const requestType: IRequestType = {
+  FAILURE: "stars/FAILURE",
+  PENDING: "stars/PENDING",
+  SUCCESS: "stars/SUCCESS"
 };
 
-/** Reducer */
-export function starsReducer(state: IStars = initialState, action: IStarsAction): IStars {
-  switch (action.type) {
-    case GET_REQUEST:
-      return Object.assign({}, state, {
-        isFetching: true
-      });
+export interface IStars {
+  stargazers_count: number;
+}
 
-    case GET_SUCCESS:
-      return Object.assign({}, state, {
-        count: action.payload.count,
-        isFetching: false
-      });
-
-    case GET_FAILURE:
-      return Object.assign({}, state, {
-        error: true,
-        isFetching: false,
-        message: action.payload.message
-      });
-
-    default:
-      return state;
+const initialState: IState<IStars> = {
+  isFetching: true,
+  payload: {
+    stargazers_count: -1
   }
+};
+
+export function starsReducer(state: IState<IStars> = initialState, action: IAction<IStars>): IState<IStars> {
+  // an example how you can deal with action type other than requesting types in promiseReducer
+  if (action.type === LOAD_STARS) {
+    return state;
+  }
+  return promiseReducer<IStars>(requestType, state, action);
 }
 
-/** Async Action Creator */
-export function getStars(): (dispatch: (action: IStarsAction) => IStars) => Promise<IStars> {
-  return (dispatch) => {
-    dispatch(starsRequest());
-
-    return fetch("https://api.github.com/repos/barbar/vortigern")
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-            .then((repo) => dispatch(starsSuccess(repo.stargazers_count)));
-        } else {
-          return res.json()
-            .then((starsErr) => dispatch(starsFailure(starsErr)));
-        }
-      })
-      .catch((err) => dispatch(starsFailure(err)));
-  };
-}
-
-/** Action Creator */
-export function starsRequest(): IStarsAction {
+export function loadStars(): IAction<IStars> {
   return {
-    type: GET_REQUEST
-  };
-}
-
-/** Action Creator */
-export function starsSuccess(count: number): IStarsAction {
-  return {
-    payload: {
-      count
-    },
-    type: GET_SUCCESS
-  };
-}
-
-/** Action Creator */
-export function starsFailure(message: any): IStarsAction {
-  return {
-    payload: {
-      message
-    },
-    type: GET_FAILURE
+    type: LOAD_STARS
   };
 }
